Add files via upload
create class User and Users, which allows you to add user and save it in the configure_file named "conf.ini". Besides, you can add mutiple secret Keys to a single user. And there are many functions as follows: show, delKeys, and so on. The functon of those is just like their names and we dont get into it here.